+++ packages/dialect-react-ui/utils/addressesUtils.ts
@@ -31,9 +31,13 @@ export const getUsersCount = (
     .filter((address) => address.address.verified)
     .filter((address) => addressTypePredicate(address.address.type));
 
-  // If there're no notifications type set up for this dapp, return all addresses count
+  // If there're no notifications type set up for this dapp, return all unique users count
   if (!subscriptions.length && notificationTypeId === null) {
-    return enabledAndVerifiedAddresses.length;
+    return new Set(
+      enabledAndVerifiedAddresses.map((addr) =>
+        addr.address.wallet.publicKey.toBase58()
+      )
+    ).size;
   }
 
   // Otherwise, filter enabled subscriptions ...
@@ -56,8 +60,7 @@ export const getUsersCount = (
     )
   );
 
-  return [...new Set(filtered.map((it) => it.wallet.publicKey.toBase58()))]
-    .length;
+  return new Set(filtered.map((it) => it.wallet.publicKey.toBase58())).size;
 };
 
 export const getAddressesCounts = (